ringbuf: Allow poll() to block for an indefinite time period
ring_buffer__poll in libbpf internally calls epoll_wait. Specifying a
timeout of -1 to epoll_wait causes it to block indefinitely if no event
transpires. We should support this functionality in libbpf-rs as well.

Currently in libbpf-rs, ringbuf's poll only takes in a Duration for the
timeout, which therefore cannot be negative. In this change, poll is
modified to take in an Option. If the user passes in None for the
timeout, then this means that poll() should block indefinitely until an
event occurs.

Signed-off-by: Joanne Koong <joannekoong@gmail.com>